Add support for external libraries within tree.
Add support to CMake for adding in external libraries into the /src/lib/libs
directory, for in tree building. To add an external library, add the
subdirectory to the /src/lib/libs directory, and change the CMakeLists.txt file
to add the subdirectory, and also add on the source list onto the LIBS_SOURCES
variable, which gets passed up the chain to be built.