Add the jansson json library as an external library.
I'll be using json as the file format used for storing things such as level
data, tile information, animation and character information, and all other
textual data files. JSON is a nice, lightweight human readable text format
which is easy to edit and work with, though choosing this library took quite
some time.

It came down to this, yajl, and cjson, I also looked at some various
c++ ones, but there weren't many. I ended up choosing this one due to it's
popularity, API, and rate of development.

The version commited is a slightly
modified version 2.3, which was released on 2012-01-27. Only modifications
being to modify the #include statements to be able to compile with CMake as
in-tree sources.

I also looked at yaml and xml as file formats, I was initially
going to go with XML with pugixml serving as the parser, but the verbosity of
XML wasn't something that I was looking forward to. YAML had various
complexities with the file format, with many things added to make it human
readable but harder to parse. I chose json due to the fact that it had a nice,
simple, rigid file format, which is easy to interface with classes and such,
and is actually commonly used as a data serialization format for transfer over
a network, such as the internet.

The very basic syntax makes it easy to edit and understand,
with little in the way of ambiguity, though one of it's strengths, having much
support for many various types of languages, was also a
weakness, due to the sheer number of libraries just for c and c++ alone, so it
took quite some time to look at various sets of information online, on what
libraries there were, their API's, strengths and weaknesses with regards to
memory usage and speed and such, rate of development, and many other aspects,
which is why there have been few commits over the past few days, in hindsight,
this took far too long.