Add logging functionality to the main header.
Added logging functionality to the /src/lib/include/impact.hh file, which
specifies 5 logging levels, CRIT, ERR, WARN, INFO, and DBG, which when used
using Log_ and the log level, will pass options to Log_OUT, specifying the
level, file, line number, and message. Currently it is set for all log levels
except for CRIT to be disabled when building with CMake using release mode,
will possibly be changed in the future with each level being configurable, but
for now, it will work. Uses fprintf for the logging functionality, due to it
already being included with SDL, and will automatically go to stdout when using
the levels INFO, WARN, DBG, and stderr when using CRIT and ERR.