Made a failed initialization exit the program.
I changed the main function a bit to make it exit out when the initialization
fails, rather than just continuing with a doomed execution. I might later
change it so that a failed sound initialization will not cause the rest of the
program to fail, as it can be a common occurance for the sound device to not be
available.