Move impact binary to root of build directory.
Before this, due to the way that I had the cmake files set up, it would put the
impact binary to src/bin of the build directory, which made it slightly harder
to access, and made testing just a little more cumbersome. So now it will put
it into the root.