Add the ImpactLoader class, which loads the config
Here is the beginnings of the ImpactLoader class, which loads files contained
within the data directory which have the tilesets, the levels and their
chapters, animation information, and all other textual data that needs to be
loaded by the game in order to make it playable. The file format is json, I'll
probably will still be working on the structure of the files, so this file and
the files in data will likely be in flux as I figure out the best way to
structure the data, and I add in more capabilities as time goes on.

For now it loads the config file, which has the list of files that list the
chapters and tilesets, those lists, and the chapter, which is a file consisting
of 1 or more levels, a set tileset, and any other information for that chapter.