Change ImpactPlay to use ImpactLoader.
I have added ImpactLoader to the ImpactPlay class, and made it so that during
the initialization function, it will load the config, load the first chapter,
and loading the first level of that chapter into the tilemap. I also added an
EntityTilemap to it as well. For now it will give me a start to be able to
begin working on rendering the level, and basic gameplay.