Add install capability to CMake.
CMake can now install the files using make install on linux, and possibly other
platforms. The default installation directory is the install directory within
the directory that you build the project. Also data files should automatically
be installed as well, if you have the submodule checked out. You can change the
default install prefix by passing -DCMAKE_INSTALL_PREFIX=/foo, only cavaet is
that currently, Impact needs the data files to be in the same directory that
the binary is installed in. I'll probably get around to changing it to be able
to be loaded from any directory, but that is the way it works for now.