Game licensed GPLv2+ and CC BY-SA, misc fixups.
Added license text to all source files, which is GPLv2+, as well as licensed
the data files as Creative Commons Attribution-ShareAlike 3.0, which allows for
commercial or noncommercial use and for derivatives, but they must be under the
same license. If I use files that are licensed some other way, I'll note it
with a license text within the directory and a mention of it in the main
LICENSE file. Also I did a few fixups of some files while I was in there,
mainly fixing some warnings with the logging text and some other annoyances.