Add a file signature define generated by lz4comp.
This took quite a while to figure out, why using just the libraries I wasn't
able to load the files generated by lz4comp. I finally realized that it stuck a
magic number at the beginning of the file to signify that it was generated by
that compressor, rather than removing it, I kept it, and implemented the same
checking within the loader to have an extra level of sanity checking.