Change loader to support lz4 compressed files
The bulk of the changes for this branch. The new file_load function of the
loader now looks by default for the json file, as always, if it finds it, it
loads it happily, if not, it looks for the compressed .lz4 file, which will be
the case for the released game, which will only have the lz4 files, but if the
level editor is used, it will generate .json files, which will, again, be
loaded over the .lz4 files. This also aids in rapid changes to the levels, as I
can run the game from the build tree, using the level editor to change the
levels, which will be put into the json files, and when I am ready for a
release, I can just run make to recompress the files, and make install to have
a ready release directory.

By default, as of this commit, the maximum supported
size for configuration files is 512K, with the maximum size for compressed
files being 128K, this will probably be good for now, but will likely change as
time goes on. I decided to have already allocated pieces of memory within the
loader for the parsing because I probably will be loading lots of little files
for each piece of textual data, and constantly allocating and deallocating what
essentially is scratch space is pretty inefficient. The memory usage used
should be minimal on most systems, and it can always be changed in the code for
later.