Change play state init function for new loader
Now that the loader can load between two different files, there is no need for
the file extension to be given along with the file name, as it chooses itself
which extension to load. Likely this will change later on down the line, as I
have some more ideas of how to structure the whole data directory, but it will
work for now for testing.