Remove loader to prepare for new ImpactIO helper
Changing around some of my plans a bit, removing ImpactLoader, which was fairly
monolithic and would have gotten to be a pretty big class, and I am going to
replace it with a generic set of functions within their own namespace,
ImpactIO, which will handle reading and writing of the .json and .lz4 files,
and also is used for a new lz4comp program that I wrote to replace the current
one, which I did not write, and didn't have the features that I wanted. This
new one utilizes the same backend as the rest of the loading functions do, so
it should remain compatible should I decide to change around the file format.
Also reading and writing of various different types of settings will be handled
within their own classes, to ensure maximum modularity.