Add new ImpactIO functions and lz4comp program
Add in the new ImpactIO functions, along with a new compressor. The ImpactIO
functions are within their own namespace to split them off from the rest of the
program. It has a number of different functions, the ones that will be used the
most are the conf_load and conf_write functions, which take in a filename, and
will either return or take in a json_t object, respectively, for the load and
write functions. It also has a number of helper functions, to parse json files,
and to do compression and decompression of lz4 files. The last 2 are used in a
new lz4comp program which should remain in parity with the rest of the lz4
handling code, as it uses the same functions. This should hopefully make it
easier in case I want to do any changes to the file format. The 2 functions
take in a character array and a length of the array, and return with a
character array, and change the length variable passed to it with the new
length. It handles it's own memory management so it should be seamless as
possible to use.