Separate lz4 functions from the rest
A slight change to make the compilation of lz4comp a bit cleaner, I seperated
out the lz4 compress and uncompress functions to their own header and source
file. They are still a part of the same namespace, and it is included as a part
of the other functions, but now they can be specified standalone from the rest
of ImpactIO functions. This enables me to not have to include all of the
unnecessary jansson libraries within the lz4comp executable. Also I forgot to
add the source files for the ImpactIO functions to the cmake build system, so I
have added them as well.