Include new config module for Impact
Include a new folder and CMakeLists.txt file for a new module of classes, which
will primarily be focused on handling the reading and writing of various forms
of textual data, and importing that data into the game itself. This data is
things such as a global configuration, tilesets, chapters, games, animation,
etc. To start off with, I'll be working on a global configuration file, which
gets written to data/config.json, which will have names for each of the
different folders, such as for images or sounds or whatever.