Add ConfigGlobal class and update data for support
This is the beginnings of the ConfigGlobal class, which handles loading and
parsing the global configuration settings. Eventually this will be passed down
along the chain and into the core state manager, which can use the current
settings it has in it for setting things such as the directories for certain
classes of files, and perhaps other settings. Currently it just creates the
object within the main function and loads the configuration. This will likely
change as time goes on as I change the structure of the global configuration,
adding or removing settings as needed.