Fix incorrect freeing of memory in loaders
I was under a misconception that delete[] was used for character arrays as well
as classes, but that is only valid when using the new keyword, as these
functions were relying on C libraries, the correct course of action is to run a
free on the data. Also I can't use regular memory management functions with
jansson, as it has it's own memory management. The correct way to do it is to
change it to use decref, as it has reference counting, and once that reaches
zero, it'll free the memory itself. So that is what it now does.