Change global config to be available to core state
The global configuration is now loaded during the instantiation of the Impact
object, which stores the state manager and various settings that need to be
preserved between state changes. Also, when adding a state onto the stack, it
will give an address of the global configuration to the state added. This will
ensure that it will be available to lower states.