Correct logger function due to multiple lines
The logger, as it was implemented, used a series of 2 printf statements, one to
actually print the line, the other to print a new line, this obviously would
screw up the functionality of single command if statements, as it would still
print the new line even if the if statement evaluated false. By wrapping it up
within a do/while block, it ensures that it is treated as one function.