Refine license text and add vim mode line
I have taken the opportunity to refine the license text located in each source
file a bit, removing a few lines that really aren't needed, and I also added
vim mode lines with the indentation and other settings that I use to edit the
files. I also took the opportunity to homogenize a couple of the files that had
slightly different formatting than the rest of the others, mainly centered
around the header files.