Make lz4comp use C rather than C++ file functions
A slight change to the way I implement the file reading and writing
functionality on lz4comp, I noticed when compiling with my cross compiler for
the pandora, that the file sizes for the programs were pretty large, as the
cross compiler I use is newer than what is on the pandora, so it was pulling in
the entire iostream framework statically, which made the file size of the
binaries pretty large in comparison. I already have the C file reading and
writing functions included due to using SDL, so I figured I may as well make
use of it, rather than needlessly pulling in libraries that I only will use for
one purpose.