Fix cross compilation with lz4comp native build
As mentioned within the commit within the data repository, this is a slight
change to the build process, creating a cmake cache file for use in any cross
compiling that needs to take place, which points to the location of the lz4comp
file for the data file compression process. Due to some issues with the way
that it worked, lz4comp now gets placed within the root of the build directory,
rather than within the data directory, the build process still works the same
though, so it's mostly a cosmetic change.