Update LZ4 to latest version
Updating the lz4 library to the latest version from subversion, revision 58.
Has a few speed improvements it looks like, and also I made minimal changes to
the files this time, moving the defines for the archive format information to
the impact_io_lz4.hh file, also there is now a function which can calculate the
largest possible size used by lz4 for the compression, and I now use it.