Add special json parsing to lz4comp
I have changed lz4comp to be able to compress json text files to a better
extent by encoding it again with jansson with the JSON_COMPACT flag on, which
will make the json data with the minimal amount of spaces, and it is without
the JSON_INDENT flag, which adds newlines between each object and array entry.
Right now for the files it doesn't do much, just a few bytes, but once I start
dealing with level data and other related data, the savings will be pretty
substantial, due to not needing many spaces and new lines. When it decompresses
it, it will add 1 space indents between the entries and other visual touches.
Only 1 space of indents as it still makes it have some structure, but reduces
the filesize on disk.