Add support for new assets and games directories
The data directory now has 2 new directories, which are assets and games,
assets contains asset lists that are used by impact for file loading and
decoupling the filenames from the rest of the file data, and also to enable
support for files to be loaded just once into memory. Games stores the
locations of the asset lists as well as any game specific settings such as
input settings, default state, etcetra.