Add orientation functionality to image_load
The image_load function of the base display state now has the ability to do
flipping and mirroring of the resulting loaded image, when passed a boolean of
whether to load multiple images or not. It is now a protected function, only
usable within the class and child classes. There are now 2 new functions,
image_multiple_load and image_single_load, which handle passing the boolean to
the image load function. If the boolean is false, only the image is loaded,
with no mirroring or flipping, with the rest of the pointers within the
DisplayAsset being NULL and having it's multiple flag set to false. If it's set
to true, it will load the image, then will mirror, flip, and mirror flip the
image, creating new surfaces and setting them to their specified pointers
within the DisplayAsset struct.