Modify ConfigGlobal to not use placeholder pointers
I didn't like the way that I implemented ConfigGlobal's loading function, using
placeholder character array pointers to hold an intermediate step for setting
the strings seemed fairly hacky to me, at best, due to the way that json
handles strings specifically within the unpack function, which is to set a
pointer to the string contained within the internal data structure, meaning
that when you have all of that overhead just for storing the strings, when I
really wanted to just use strings and free the json object when I am done with
it. I am doing the same sort of thing within the code that I did before, which
is the pointers are used to set the strings, but now I don't have that
intermediary step, and just set the strings directly from the json data.