Add new ConfigBase class, and add it to CMake
There is a new ConfigBase class, which will be the base class for all
configuration parsing classes. Currently it just has a function to handle
string parsing from the json object. I decided to do this to have a single
unified way of getting strings from the json data, and the function is able to
parse out strings from objects and arrays, as well as just directly getting the
string data from a json_t object containing just a string. It also has some
error handling, returning an empty string on error, and will also do basic
error reporting for the following cases:

JSON object passed is NULL
Object or Array index is not present
Object or Array index is present but is not a string
Unsupported type passed