Change input to set bit for first frame pressed
Here is a slight change to the input handling code, which makes it also set a
bit for the frame that the button was set, which resets on the next frame. This
will enable things like setting a status for a button that was just pressed, or
any number of other behaviors that you want to only occur when you first press
a button.