Fix SDL includes to increase portability
SDL requires you to just include the SDL.h header separate, without the
preceding SDL/, due to the nature of some build systems having SDL within the
root of the include directory. This is a small fix to do that, adding the
include directories detected by CMake to the include paths, which should ensure
that it is cross platform capable.