Fix jansson to compile without C99 support
The visual studio C compiler doesn't have support for C99, so the inline
keyword does not exist. This is one of the first commits that I'll eventually
do to try to get it fully building within visual studio, but for now I'll
continue developing the game, and it can be built with MinGW for now anyway.