Fix binaries with correct usage of SDL_main
The way that SDL works is that it wraps up the main function within it's own
SDL_main function. For linux and other unix operating systems, it will link
against SDL_main automatically, and correctly set all the settings. But when
building on other platforms, such as windows and Mac OSX, it requires special
consideration for the use, both with cmake as well as the functions themselves.
In the case of CMake, I had read some incorrect documentation that said to add
SDLmain to the link libraries setting, when in fact it was already added with
SDL_LIBRARIES, and it was not needed. So adding it was redundant, and windows
could not find SDLmain.lib, so it would throw a link error. In the case of the
binaries, it is best to include SDL_main.h, and also have the correct main
signature in order for it to work. So these changes were added in order to get
it compiling using visual studio. There are still issues relating to the
jansson json library which I will fix, as well as even after fixing it to
compile, it wasn't parsing correctly, so I'll have to look into that.