Provide a variable for CMake to point to SDL Libs
If you are compiling on a platform that doesn't have a set development path
that get looked at for development headers and libraries to get included, such
as Windows, this setting will provide a path for the FindSDL.cmake module to
find, and it will only get set if the user sets it manually. For compiling on
windows, you will copy the include and lib directories from the SDL devel
packages, as well as the modules, and stick them all in the same folder. Making
sure to move the libraries for your platform from their directory to the root
of the lib directory. For example, moving all the libraries from x86 to the
root of lib, if building for a 32 bit OS. When you configure it now using the
cmake gui, giving it this directory, all the variables should be set, with no
NOT_FOUND's being listed.