Provide snprintf and strtoll for jansson for VS
Visual studio doesn't support C99, so the two functions used by jansson strtoll
and snprintf do not exist. In order to get it to compile under visual studio,
you have to give it the visual studio equivelants, _strtoi64 and _snprintf,
respectively. By doing this, you can get it to compile, though functionality
for each may not be 1:1, but it's close enough to work. The ifdef will ensure
that it is only included under visual studio.