Fix lz4comp on windows and perhaps others
The implementation of the json file loading and writing was slightly different
to my implmentation within ImpactIO, which made it work incorrectly under
windows. These few lines ensure that the string itself is NULL terminated,
which makes jansson able to parse the files correctly. Before it was giving
errors trying to parse the last byte, now it parses it correctly, and was able
to compress and uncompress multiple times without issue.