Make Jansson VS handling cleaner
A slight change to the way that I handle the visual studio detection, having
the detection for visual studio being a branch of the checks rather than within
the checks, as I noticed the way that I implemented it before, I wasn't getting
inlining when compiling under linux with gcc. Now it should.